Add addHostileRoom function and relocate code
