+++ src/modules/flagsManagement.ts
@@ -1,5 +1,4 @@
-import { addColonizationOperation } from './empireManagement';
-import { unclaimRoom } from './memoryManagement';
+import { addColonizationOperation, addHostileRoom, unclaimRoom } from './empireManagement';
 
 export default function manageFlags() {
     if (Game.flags.colonize) {
@@ -10,5 +9,10 @@ export default function manageFlags() {
     if (Game.flags.unclaim) {
         unclaimRoom(Game.flags.unclaim.pos.roomName);
         Game.flags.unclaim.remove();
+    }
+
+    if (Game.flags.hostile) {
+        addHostileRoom(Game.flags.hostile.pos.roomName);
+        Game.flags.hostile.remove();
     }
 }
