+++ src/modules/memoryManagement.ts
@@ -25,22 +25,3 @@ function handleDeadCreep(creepName: string) {
 
     delete Memory.creeps[creepName];
 }
-
-export function unclaimRoom(roomName: string) {
-    let room = Game.rooms[roomName];
-
-    if (room?.controller?.my) {
-        room.controller.unclaim();
-    }
-
-    if (room?.find(FIND_MY_CONSTRUCTION_SITES).length) {
-        room.find(FIND_MY_CONSTRUCTION_SITES).forEach((site) => site.remove());
-    }
-
-    Memory.empire.colonizationOperations = Memory.empire.colonizationOperations.filter((op) => op.destination !== roomName);
-    Memory.empire.spawnAssignments = Memory.empire.spawnAssignments.filter((asssignment) => asssignment.designee !== roomName);
-
-    delete Memory.rooms[roomName];
-
-    return 'done';
-}
